chore(rename): llmrouter -> llm_router across package, imports, docs, tests
