+++ tests/test_monitor_resources.py
@@ -1,4 +1,4 @@
-from llmrouter.monitor_resources import start_resource_monitoring
+from llm_router.monitor_resources import start_resource_monitoring
 
 
 def test_monitor_thread_starts(tmp_path, monkeypatch):
